Add support for Twitter Cards and Open Graph
Twitter Cards and Facebook's Open Graph are implemented with special
meta tags and quite widely used for embedding. Checking for both of them
is both the official recommendation in the Twitter docs and part of
Slack's strategy for what they call "unfurling".

https://api.slack.com/docs/unfurling

This is basic support and doesn't handle OEmbed or media embeds (like
videos, music, or plain images), but that shouldn't be that hard. -POLM